added_posts

+++ api/v1/controller/posts.py
@@ -6,12 +6,13 @@ from sqlalchemy import *
 from sqlalchemy.orm import *
 from flask import abort, jsonify, make_response, request
 from api.v1.controller.path import app_controller
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
 from models import *
 
 def serialized_posts(post):
     """Serialize a User object to a JSON-serializable format."""
     return {
-        'id': post.id,
+        'post_id': post.postid,
         'title': post.title,
         'user_id': post.user_id,
         'filepath': post.filepath,
@@ -25,45 +26,65 @@ def serialized_posts(post):
 
 @app_controller.route('/users/<user_id>/posts', methods=['GET'],
                  strict_slashes=False)
+# @jwt_required()
 def get_posts_by_userId(user_id):
     """
     Retrieves the list of all posts objects
     of a specific User, or a specific user
     """
-    list_cities = []
-    user = User.query.filter_by(id=user_id).first()
+#    current_user = get_jwt_identity()
+#    if not current_user:
+#        return jsonify({"message": "Not a valid user"}), 401
+#    user = User.query.filter_by(email=current_user).first()
+#    if not user:
+#        return jsonify({"message": "Not a valid user"}), 401
+
+    list_posts = []
+    user = User.query.filter_by(userid=user_id).first()
     if not user:
         abort(404)
     for post in user.posts:
-        list_cities.append(serialized_posts(post))
-    return jsonify(list_cities)
+        list_posts.append(serialized_posts(post))
+    return jsonify(list_posts)
 
 @app_controller.route('/posts', methods=['GET'], strict_slashes=False)
+# @jwt_required()
 def get_posts():
-    """Create a new view for post object that handles
-    all default RESTFul API actions:"""
-    # users = []
+    """get all posts"""
+    # current_user = get_jwt_identity()
+    # if not current_user:
+    #    return jsonify({"message": "Not a valid user"}), 401
+    # user = User.query.filter_by(email=current_user).first()
+    # if not user:
+    #     return jsonify({"message": "Not a valid user"}), 401
+
     posts = Post.query.order_by(Post.created.desc())
     serialized_post = [serialized_posts(post) for post in posts]
     return jsonify(serialized_post)
 
-@app_controller.route('/posts/<int:id>', methods=['GET'], strict_slashes=False)
+@app_controller.route('/posts/<id>', methods=['GET'], strict_slashes=False)
 def get_post_by_id(id):
-    """Create a new view for post object that handles
-    all default RESTFul API actions:"""
+    """get post by id"""
     # users = []
-    post = Post.query.filter_by(id=id).first()
+    post = Post.query.filter_by(postid=id).first()
     if post is None:
         abort(404)
     serialized_post = serialized_posts(post)
     return jsonify(serialized_post)
 
-@app_controller.route('/posts/<int:id>', methods=['DELETE'],
+@app_controller.route('/posts/<id>', methods=['DELETE'],
                  strict_slashes=False)
+@jwt_required()
 def delete_post(id):
-    """Create a new view for User object that
-    handles all default RESTFul API actions:"""
-    post = Post.query.filter_by(id=id).first()
+    """Delete a post by id"""
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+
+    post = Post.query.filter_by(postid=id).first()
     if post is None:
         abort(404)
     db.session.delete(post)
@@ -71,10 +92,17 @@ def delete_post(id):
     return (jsonify({}))
 
 @app_controller.route('/users/<user_id>/posts', methods=['POST'], strict_slashes=False)
+@jwt_required()
 def post_blog(user_id):
-    """Create a new view for User object
-    that handles all default RESTFul API actions:"""
-    user = User.query.filter_by(id=user_id).first()
+    """Create a post for a specific user"""
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+
+    user = User.query.filter_by(userid=user_id).first()
     if not user:
         abort(404)
     if not request.get_json():
@@ -84,17 +112,24 @@ def post_blog(user_id):
     if 'body' not in request.get_json():
         return make_response(jsonify({'error': 'Missing body'}), 400)
     post = Post(**request.get_json())
-    post.user_id = user.id
+    post.user_id = user.userid
     db.session.add(post)
     db.session.commit()
     return make_response(jsonify(serialized_posts(post)), 201)
 
-@app_controller.route('/posts/<int:id>', methods=['PUT'],
+@app_controller.route('/posts/<id>', methods=['PUT'],
                  strict_slashes=False)
+@jwt_required()
 def put_post(id):
-    """Create a new view for User object that
-    handles all default RESTFul API actions:"""
-    post = Post.query.filter_by(id=id).first()
+    """update post """
+    current_user = get_jwt_identity()
+    if not current_user:
+        return jsonify({"message": "Not a valid user"}), 401
+    user = User.query.filter_by(email=current_user).first()
+    if not user:
+        return jsonify({"message": "Not a valid user"}), 401
+
+    post = Post.query.filter_by(postid=id).first()
     if post is None:
         abort(404)
     if not request.get_json():
